change check null and create XBottomSheet
